+++ internal/streamingcoord/server/balancer/channel/pchannel.go
@@ -36,11 +36,6 @@ type PChannelMeta struct {
 	inner *streamingpb.PChannelMeta
 }
 
-// IsRWChannel check if the pchannel is a rw channel.
-func (c *PChannelMeta) IsRWChannel() bool {
-	return c.inner.Channel.GetAccessMode() == streamingpb.PChannelAccessMode_PCHANNEL_ACCESS_READWRITE
-}
-
 // Name returns the name of the channel.
 func (c *PChannelMeta) Name() string {
 	return c.inner.GetChannel().GetName()
