--- internal/streamingcoord/server/balancer/policy/builder.go
@@ -1,1 +1,0 @@
-package policy
