+++ internal/streamingcoord/server/balancer/policy/vchannelfair/pchannel_affinity.go
@@ -1,17 +1,23 @@
-package policy
+package vchannelfair
 
 import (
 	"github.com/milvus-io/milvus/internal/streamingcoord/server/balancer/channel"
 	"github.com/milvus-io/milvus/pkg/v2/streaming/util/types"
 )
 
+// newPChannelAffinity creates a new pchannel affinity from the given channels.
 func newPChannelAffinity(channels map[types.ChannelID]channel.PChannelStatsView) *pchannelAffinity {
 	affinities := make(map[sortedChannelID]float64)
 	for channelID1, stats1 := range channels {
 		for channelID2, stats2 := range channels {
+			id := newSortedChannelID(channelID1, channelID2)
 			if channelID1 == channelID2 {
 				continue
 			}
+			if _, ok := affinities[id]; ok {
+				continue
+			}
+
 			counter := map[int64]int{}
 			for _, collectionID := range stats1.VChannels {
 				counter[collectionID]++
@@ -29,7 +35,7 @@ func newPChannelAffinity(channels map[types.ChannelID]channel.PChannelStatsView)
 			if len(counter) > 0 {
 				affinity = float64(len(counter)-repeated) / float64(len(counter))
 			}
-			affinities[newSortedChannelID(channelID1, channelID2)] = affinity
+			affinities[id] = affinity
 		}
 	}
 	return &pchannelAffinity{
@@ -38,25 +44,31 @@ func newPChannelAffinity(channels map[types.ChannelID]channel.PChannelStatsView)
 }
 
 // pchannelAffinity is a affinity relationship between pchannel and pchannel.
+// It describe how much two pchannels can be assigned to same streamingnode without any hotspot at collection level.
+// If there're no repeated collection between two pchannels, the affinity is 1.
+// If there're all repeated collections between two pchannels, the affinity is 0.
 type pchannelAffinity struct {
 	affinity map[sortedChannelID]float64 // range from 0 to 1, 1 means high affinity.
 }
 
-// GetAffinity returns the affinity between two pchannels.
-func (a *pchannelAffinity) GetAffinity(channelID1 types.ChannelID, channelID2 types.ChannelID) float64 {
+// MustGetAffinity returns the affinity between two pchannels.
+func (a *pchannelAffinity) MustGetAffinity(channelID1 types.ChannelID, channelID2 types.ChannelID) float64 {
 	sortedID := newSortedChannelID(channelID1, channelID2)
 	affinity, ok := a.affinity[sortedID]
 	if !ok {
-		return 0
+		panic("affinity not found")
 	}
 	return affinity
 }
 
+// sortedChannelID is a sorted channel ID.
+// It is used to avoid duplicate key in the affinity map.
 type sortedChannelID struct {
 	channelID1 types.ChannelID
 	channelID2 types.ChannelID
 }
 
+// newSortedChannelID creates a new sorted channel ID.
 func newSortedChannelID(id1, id2 types.ChannelID) sortedChannelID {
 	if id1.LT(id2) {
 		return sortedChannelID{
